requirements.txt
Flask==1.1.1
google-api-core==1.16.0
google-auth==1.11.0
google-cloud-vision==0.38.1
googleapis-common-protos==1.51.0
requests==2.22.0
pyyaml
requests
pillow